+++ animation_nodes/base_types/nodes/base_node.py
@@ -52,6 +52,7 @@ class AnimationNode:
     activeOutputIndex: IntProperty()
 
     searchTags = []
+    isSearchable = True
     onlySearchTags = False
 
     # can contain: 'NO_EXECUTION', 'NOT_IN_SUBPROGRAM',