+++ animation_nodes/nodes/subprogram/loop_generator_output.py
@@ -10,6 +10,7 @@ class LoopGeneratorOutputNode(bpy.types.Node, AnimationNode):
     bl_idname = "an_LoopGeneratorOutputNode"
     bl_label = "Loop Generator Output"
     dynamicLabelType = "ALWAYS"
+    isSearchable = False
 
     def settingChanged(self, context):
         self.refresh()